requirements.txt
pytest==7.4.4
allure-pytest==2.13.2
selene==2.0.0rc4
python-dotenv==1.0.0
requests~=2.31.0
selenium~=4.18.1

requests
curlify
jsonschema